requirements.txt
annotated-types==0.5.0
anyio==3.7.1
astor==0.8.1
attrdict==2.0.1
Babel==2.14.0
bce-python-sdk==0.9.21
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==3.0.11
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.103.2
fire==0.6.0
Flask==2.2.5
flask-babel==3.1.0
fonttools==4.38.0
future==1.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.8
imageio==2.31.2
imgaug==0.4.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
kiwisolver==1.4.5
lmdb==1.5.1
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
openpyxl==3.1.3
opt-einsum==3.3.0
packaging==24.0
paddleocr==2.7.0.2
paddlepaddle==2.5.2
pandas==1.3.5
pdf2docx==0.5.8
Pillow==9.5.0
premailer==3.10.0
protobuf==3.20.2
psutil==6.0.0
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
PyMuPDF==1.20.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
PyWavelets==1.3.0
PyYAML==6.0.1
rapidfuzz==3.4.0
rarfile==4.2
requests==2.31.0
scikit-image==0.19.3
scipy==1.7.3
setuptools==68.0.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
starlette==0.27.0
termcolor==2.3.0
tifffile==2021.11.2
tqdm==4.66.5
typing_extensions==4.7.1
urllib3==2.0.7
visualdl==2.5.3
Werkzeug==2.2.3
wheel==0.41.2
zipp==3.15.0
